Update 34.md
Today was challenge day, where I had to create multiple animations for the flag game app.  The problem was I wanted the alert to trigger after the animation, so I had to use the dispatch queue to set a timer to have the code run after the animation.  This took me two hours and 20 minutes to complete.